# api/docs/swagger.yaml
basePath: /v1
definitions:
  http.Response:
    properties:
      data: {}
      description:
        type: string
      status:
        type: string
    type: object
...
  models.ClinicReq:
    properties:
      address:
        type: string
      image_link:
        type: string
      name:
        type: string
      phone_number:
        type: string
    type: object
  models.ClinicRes:
    properties:
      address:
        type: string
      id:
        type: integer
      image_link:
        type: string
      name:
        type: string
      phone_number:
        type: string
    type: object
...
  models.ResponseMessage:
    properties:
      message:
        type: string
    type: object
...
  models.UpdateClinicReq:
    properties:
      address:
        type: string
      id:
        type: integer
      image_link:
        type: string
      name:
        type: string
      phone_number:
        type: string
    type: object
...
info:
  contact: {}
  description: This is Clinic server api
  termsOfService: 1 term Traffic Light
  title: Clinic
  version: "1.0"
paths:
...
  /clinic:
    post:
      consumes:
      - application/json
      description: Create Clinic
      operationId: createclinic
      parameters:
      - description: CreateClinicInfo
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.ClinicReq'
      produces:
      - application/json
      responses:
        "201":
          description: Clinic data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.ClinicRes'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Create Clinic
      tags:
      - Clinic
  /clinic/{id}:
    get:
      consumes:
      - application/json
      description: GET Clinic
      operationId: getclinic
      parameters:
      - description: GetClinicInfo
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Clinic data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.ClinicRes'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: GET Clinic
      tags:
      - Clinic
  /clinics:
    get:
      consumes:
      - application/json
      description: GET Clinics
      operationId: getclinics
      produces:
      - application/json
      responses:
        "201":
          description: Clinics data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.ClinicRes'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: GET Clinics
      tags:
      - Clinic
    put:
      consumes:
      - application/json
      description: PUT Clinics
      operationId: putclinic
      parameters:
      - description: UpdateClinics
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.UpdateClinicReq'
      produces:
      - application/json
      responses:
        "200":
          description: Client data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.ClinicRes'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: PUT Clinics
      tags:
      - Clinic
  /clinics/delete/{id}:
    delete:
      consumes:
      - application/json
      description: PUT Clinics
      operationId: deleteclinics
      parameters:
      - description: clinicID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Clinic data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.ResponseMessage'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: PUT Clinics
      tags:
      - Clinic
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"